🤖 Refactor(#21): modify component name and path of icons
