+++ src/components/common-components/common/Footer.tsx
@@ -14,7 +14,7 @@ const Footer = () => {
     <footer className="w-full h-[264px] bottom-0 flex flex-col flex-wrap justify-center content-center items-center text-gray-06 bg-gray-03">
       <div className={variants.title}>
         <Image
-          src="/assets/ddoba_logo_black-footer.svg"
+          src="/assets/components/ddoba_logo_black-footer.svg"
           alt="logo"
           width={32}
           height={26}
